+++ exolve-from-ipuz.js
@@ -24,7 +24,7 @@ SOFTWARE.
 The latest code and documentation for Exolve can be found at:
 https://github.com/viresh-ratnakar/exolve
 
-Version: Exolve v1.25 October 31 2021
+Version: Exolve v1.26 November 3 2021
 */
 
 /**